refactorizacion de clases/packages
