+++ src/main/java/model/Room.java
@@ -1,6 +1,6 @@
-package models;
+package model;
 
-import exceptions.*;
+import exception.*;
 import java.util.ArrayList;
 import java.util.List;
